requirements.txt
python-gnupg
termcolor
